Рефакторинг: перенёс логику методов в repository, добавил метод getStorage для получения данных из хранилища, исправил баг со сбросом старого пароля на новый
